Update decorators to stage-2 syntax
Memoized was written using a new and unstable syntax for class decorators. The
proposal for decorators has changed a bit since then, but is thankfully quite a
bit more stable as it gets closer to becoming an official ecma standard!

See https://codeburst.io/no-need-to-wait-for-the-holidays-start-decorating-now-67b9dabd60d7
and https://github.com/tc39/proposal-decorators/